auditorias de las 3 entidades funcionando

+++ src/service/RemitoService.ts
@@ -5,6 +5,10 @@ import ProductoRepository from "../repository/ProductoRepository";
 import RemitoDto from "../model/DTO/RemitoDto";
 import EventoRepository from "../repository/EventoRepository";
 import EventoService from "./EventoService";
+import AuditoriaRemitoService from "./AuditoriaRemitoService";
+import UsuarioRepository from "../repository/UsuarioRepository";
+import AuditoriaStockService from "./AuditoriaStockService";
+
 
 const listarRemitos = async () => {
   return await RemitoRepository.findall();
@@ -29,27 +33,44 @@ const buscarProducto = async (detalle: any) => {
 };
 
 
-const actualizarOCrearProducto = async (detalle: any) => {
+const actualizarOCrearProducto = async (detalle: any, firebaseUid: string) => {
+  const usuario = await UsuarioRepository.findByFirebaseUid(firebaseUid);
+  if (!usuario || !usuario._id) throw new Error("Usuario no encontrado");
+
   let producto = await buscarProducto(detalle);
+  let tipoOperacion = "CREACIÓN";
+  let valorAnterior = "-";
 
   if (producto) {
-    if (!producto._id) {
-  throw new Error("El producto no tiene ID definido");
-}
-    const nuevaCantidad = producto.cantidad_actual + detalle.cantidad;
-    await ProductoRepository.update(producto._id, {cantidad_actual: nuevaCantidad,});
+    if (!producto._id) throw new Error("El producto no tiene ID definido");
 
-    detalle.id_producto = producto._id.toString();
-    detalle.nombre_producto = producto.nombre_producto;
+    const nuevaCantidad = producto.cantidad_actual + detalle.cantidad;
+    valorAnterior = JSON.stringify({ cantidad_actual: producto.cantidad_actual });
+
+    await ProductoRepository.update(producto._id, { cantidad_actual: nuevaCantidad });
+    producto.cantidad_actual = nuevaCantidad;
+    tipoOperacion = "AJUSTE";
   } else {
-    producto = await ProductoRepository.create({nombre_producto: detalle.nombre_producto,cantidad_actual: detalle.cantidad,});
-     
-    if(!producto._id){
-      throw new Error("Error al crear producoto: no se genero id")
-    }
-    detalle.id_producto = producto._id.toString();
-    detalle.nombre_producto = producto.nombre_producto;
+    producto = await ProductoRepository.create({
+      nombre_producto: detalle.nombre_producto,
+      cantidad_actual: detalle.cantidad,
+    });
+
+    if (!producto._id) throw new Error("Error al crear producto: no se generó ID");
   }
+
+  detalle.id_producto = producto._id.toString();
+  detalle.nombre_producto = producto.nombre_producto;
+
+  await AuditoriaStockService.registrarAuditoria({
+    id_stock: producto._id.toString(),
+    id_usuario: usuario._id.toString(),
+    nombre_usuario: usuario.nombre,
+    campo_modificado: tipoOperacion,
+    valor_anterior: valorAnterior,
+    valor_nuevo: JSON.stringify({nombre_producto: producto.nombre_producto,cantidad_actual: producto.cantidad_actual, }),
+    descripcion: `Producto ${producto.nombre_producto} ${tipoOperacion === "CREACIÓN" ? "creado" : "ajustado"} desde remito`,
+  });
 
   return detalle;
 };
@@ -74,16 +95,38 @@ const validarFecha = (fecha: Date) => {
   return fechas; 
 };
 
-const crearRemito = async (remitoDto: RemitoDto) => {
-remitoDto.fecha = validarFecha(remitoDto.fecha);
+export const formatearRemitoParaAuditoria = (remito: any) => {
+  const {numero_remito,fecha,empresa,productos,recibido_por,estado,} = remito;
+
+  const productosFormateados = productos.map((p: any) => ({
+    nombre_producto: p.nombre_producto,
+    cantidad: p.cantidad,
+  }));
+
+  const recibidoFormateado = recibido_por
+    ? {
+        nombre: recibido_por.nombre,
+        tipo_persona: recibido_por.tipo_persona,
+      }
+    : null;
+
+  return {numero_remito,fecha,empresa,productos: productosFormateados,recibido_por: recibidoFormateado,estado,};
+};
+
+const crearRemito = async (remitoDto: RemitoDto,firebaseUid: string) => {
+  remitoDto.fecha = validarFecha(remitoDto.fecha);
   const existe = await  RemitoRepository.findByNumero(remitoDto.numero_remito);
   if (existe){
    throw new Error(`Ya existe una factura con el número ${remitoDto.numero_remito}`);
   }
   
+
+
+
+  
   const productosActualizados = [];
   for (const detalle of remitoDto.productos) {
-    const actualizado = await actualizarOCrearProducto(detalle);
+    const actualizado = await actualizarOCrearProducto(detalle, firebaseUid);
     productosActualizados.push(actualizado);
   }
 
@@ -102,17 +145,103 @@ remitoDto.fecha = validarFecha(remitoDto.fecha);
   entidad_afectada: "Remito",
   descripcion: `Se creó un nuevo remito de ${remitoDto.empresa}`
 });
+  
+const usuario = await UsuarioRepository.findByFirebaseUid(firebaseUid);
+if (!usuario || !usuario._id) throw new Error("Usuario no encontrado");
+
+const remitoCompleto = await RemitoRepository.findById(nuevoRemito._id);
+const remitoAuditable = formatearRemitoParaAuditoria(remitoCompleto);
+
+await AuditoriaRemitoService.registrarAuditoria({
+  id_remito: nuevoRemito._id.toString(),
+  id_usuario: usuario._id.toString(),
+  numero_remito: nuevoRemito.numero_remito,
+  nombre_usuario: usuario.nombre,
+  campo_modificado: "CREACIÓN",
+  valor_anterior: "-",
+  valor_nuevo: JSON.stringify(remitoAuditable),
+  descripcion: `Se creó el remito ${nuevoRemito.numero_remito} por ${usuario.nombre}`,
+});
+
+  
   return nuevoRemito ;
 };
 
 
 
-const actualizarRemito = async (id: string, remitoDto: Partial<RemitoDto>) => {
-  return await RemitoRepository.update(id, remitoDto);
+const actualizarRemito = async (id: string, remitoDto: Partial<RemitoDto>,firebaseUid: string) => {
+ const usuario = await UsuarioRepository.findByFirebaseUid(firebaseUid);
+  if (!usuario || !usuario._id) throw new Error("Usuario no encontrado");
+
+  const remitoExistente = await RemitoRepository.findById(id);
+  if (!remitoExistente) throw new Error("Remito no encontrado");
+
+  
+  const remitoActualizado = await RemitoRepository.update(id, remitoDto);
+  if (!remitoActualizado || !remitoActualizado._id)
+    throw new Error("Error al actualizar el remito");
+  
+  const existenteAuditable = formatearRemitoParaAuditoria(remitoExistente);
+const actualizadoAuditable = formatearRemitoParaAuditoria(remitoActualizado);
+
+  const valorAnterior = {
+    fecha: remitoExistente.fecha?.toString() !== remitoActualizado.fecha?.toString()
+      ? remitoExistente.fecha
+      : "-",
+    empresa:
+      remitoExistente.empresa !== remitoActualizado.empresa
+        ? remitoExistente.empresa
+        : "-",
+  };
+
+  const valorNuevo = {
+    fecha: remitoExistente.fecha?.toString() !== remitoActualizado.fecha?.toString()
+      ? remitoActualizado.fecha
+      : "-",
+    empresa:
+      remitoExistente.empresa !== remitoActualizado.empresa
+        ? remitoActualizado.empresa
+        : "-",
+  };
+
+  if (valorAnterior.fecha !== "-" || valorAnterior.empresa !== "-") {
+    await AuditoriaRemitoService.registrarAuditoria({
+      id_remito: remitoActualizado._id.toString(),
+      id_usuario: usuario._id.toString(),
+      numero_remito: remitoActualizado.numero_remito,
+      nombre_usuario: usuario.nombre,
+      campo_modificado: "ACTUALIZACIÓN",
+      valor_anterior: JSON.stringify(valorAnterior),
+      valor_nuevo: JSON.stringify(valorNuevo),
+      descripcion: `Remito ${remitoActualizado.numero_remito} actualizado por ${usuario.nombre}`,
+    });
+  }
+
+  return remitoActualizado;
 };
 
-const borrarRemito = async (id: string) => {
-  return await RemitoRepository.remove(id);
+const borrarRemito = async (id: string,firebaseUid: string) => {
+const usuario = await UsuarioRepository.findByFirebaseUid(firebaseUid);
+if (!usuario || !usuario._id) throw new Error("Usuario no encontrado");
+
+const remitoExistente = await RemitoRepository.findById(id);
+const remitoAuditable = formatearRemitoParaAuditoria(remitoExistente);
+if (!remitoExistente) throw new Error("Remito no encontrado");
+
+await AuditoriaRemitoService.registrarAuditoria({
+  id_remito: id,
+  id_usuario: usuario._id.toString(),
+  numero_remito: remitoExistente.numero_remito,
+  nombre_usuario: usuario.nombre,
+  campo_modificado: "ELIMINACIÓN",
+  valor_anterior: JSON.stringify(remitoAuditable),
+  valor_nuevo: "-",
+  descripcion: `Remito ${remitoExistente.numero_remito} eliminado por ${usuario.nombre}`,
+});
+
+await RemitoRepository.remove(id);
+return remitoExistente;
+
 };
 
 const obtenerPorNumero = async (numero: number): Promise<RemitoDto | null> => {
